Use superclass for "candle-lit" features
